constants for everything
